Fallback to select for /dev/tty on Darwin/BSD
