+++ cancelreader_unix.go
@@ -4,16 +4,8 @@
 package tea
 
 import (
-	"errors"
-	"fmt"
 	"io"
-	"os"
-	"strings"
-
-	"golang.org/x/sys/unix"
 )
-
-var selectMaxFd = 1024
 
 // newCancelReader returns a reader and a cancel function. If the input reader
 // is an *os.File, the cancel function can be used to interrupt a blocking call
@@ -23,120 +15,5 @@ var selectMaxFd = 1024
 // returns false. The generic unix implementation is based on the posix select
 // syscall.
 func newCancelReader(reader io.Reader) (cancelReader, error) {
-	file, ok := reader.(*os.File)
-	if !ok || file.Fd() >= uintptr(selectMaxFd) {
-		return newFallbackCancelReader(reader)
-	}
-	r := &selectCancelReader{file: file}
-
-	var err error
-
-	r.cancelSignalReader, r.cancelSignalWriter, err = os.Pipe()
-	if err != nil {
-		return nil, err
-	}
-
-	return r, nil
-}
-
-type selectCancelReader struct {
-	file               *os.File
-	cancelSignalReader *os.File
-	cancelSignalWriter *os.File
-	cancelMixin
-}
-
-func (r *selectCancelReader) Read(data []byte) (int, error) {
-	if r.isCancelled() {
-		return 0, errCanceled
-	}
-
-	for {
-		err := waitForRead(r.file, r.cancelSignalReader)
-		if err != nil {
-			if errors.Is(err, unix.EINTR) {
-				continue // try again if the syscall was interrupted
-			}
-
-			if errors.Is(err, errCanceled) {
-				// remove signal from pipe
-				var b [1]byte
-				_, readErr := r.cancelSignalReader.Read(b[:])
-				if readErr != nil {
-					return 0, fmt.Errorf("reading cancel signal: %w", readErr)
-				}
-			}
-
-			return 0, err
-		}
-
-		return r.file.Read(data)
-	}
-}
-
-func (r *selectCancelReader) Cancel() bool {
-	r.setCancelled()
-
-	// send cancel signal
-	_, err := r.cancelSignalWriter.Write([]byte{'c'})
-	if err != nil {
-		return false
-	}
-
-	return true
-}
-
-func (r *selectCancelReader) Close() error {
-	var errMsgs []string
-
-	// close pipe
-	err := r.cancelSignalWriter.Close()
-	if err != nil {
-		errMsgs = append(errMsgs, fmt.Sprintf("closing cancel signal writer: %v", err))
-	}
-
-	err = r.cancelSignalReader.Close()
-	if err != nil {
-		errMsgs = append(errMsgs, fmt.Sprintf("closing cancel signal reader: %v", err))
-	}
-
-	if len(errMsgs) > 0 {
-		return fmt.Errorf(strings.Join(errMsgs, ", "))
-	}
-
-	return nil
-}
-
-func waitForRead(reader *os.File, abort *os.File) error {
-	readerFd := int(reader.Fd())
-	abortFd := int(abort.Fd())
-
-	maxFd := readerFd
-	if abortFd > maxFd {
-		maxFd = abortFd
-	}
-
-	// this is a limitation of the select syscall
-	if maxFd >= selectMaxFd {
-		return fmt.Errorf("cannot select on file descriptor %d which is larger than 1024", maxFd)
-	}
-
-	fdSet := &unix.FdSet{}
-	fdSet.Set(int(reader.Fd()))
-	fdSet.Set(int(abort.Fd()))
-
-	_, err := unix.Select(maxFd+1, fdSet, nil, nil, nil)
-	if err != nil {
-		return fmt.Errorf("select: %w", err)
-	}
-
-	if fdSet.IsSet(abortFd) {
-		return errCanceled
-	}
-
-	if fdSet.IsSet(readerFd) {
-		return nil
-	}
-
-	return fmt.Errorf("select returned without setting a file descriptor")
+	return newSelectCancelReader(reader)
 }
